"post-order" regulations

+++ Dtos/ProductDto/GetProductDto.cs
@@ -1,5 +1,6 @@
 ﻿using ESShopReact.NetCore.Data;
 using ESShopReact.NetCore.Dtos.Category;
+using ESShopReact.NetCore.Dtos.OrderDto;
 using ESShopReact.NetCore.Models;
 using System.Collections.Generic;
 
@@ -12,7 +13,7 @@ namespace ESShopReact.NetCore.Dtos.Product
         public string Description { get; set; }
         public int UnitsInStock { get; set; }
         public int OrderedQuantity { get; set; }
-        public int ListPrice { get; set; }
+        public decimal ListPrice { get; set; }
         public int CategoryID { get; set; }
         public GetCategoryDto Category { get; set; }
