+++ Dtos/ProductDto/UpdateProductDto.cs
@@ -10,7 +10,7 @@ namespace ESShopReact.NetCore.Dtos.Product
         public string Description { get; set; }
         public int UnitsInStock { get; set; }
         public int OrderedQuantity { get; set; }
-        public int ListPrice { get; set; }
+        public decimal ListPrice { get; set; }
         public int CategoryID { get; set; }
         public GetCategoryDto Category { get; set; }
     }
